Handle HasRepped constraints.
The Rep for a constraint is a dict, but there is no actual constraint held in
the tuple in this case, so we can't just look it up there. Instead, we now use
the usual `buildDictionary` approach if we fail to find the constraint in the
tuple.

This allows us to properly categorize GADTs with constraints in the
alternatives.