Eliminate extraneous builds.
Generally, the third component of a GHC version indicates
backward-compatibility. That is, GHC-8.10.2 should successfully execute any
well-defined Haskell program that worked on GHC-8.10.1. So, we only need to test
on the lowest-valued version that we support.

However, this isn't always true. E.g., 8.10.7 introduced breaking changes that
we explicitly conditionalize on, so we must also test that version.

This is intended to reduce the size of the CI matrix with minimal impact on our
coverage, which makes it easier to support more GHC versions and also add other
dimensions without hitting concurrent build limits as much.