Support GHC 9.2.
The current sticking point here is a panic in `refineFromInScope`. It seems like
we've had an issue here for a while, but until GHC 9.2, it was only a warning
and only printed when GHC was compiled in debug mode. This occurs (at least)
when we try to `simplifyExpr` on a generated `fix`, like:

> fix
>   @(Word8 -> Word8)
>   (\ (rec_Bh :: Word8 -> Word8) (x_acFz :: Word8) ->
>      case > @Word8 $fOrdWord8 x_acFz (fromInteger @Word8 $fNumWord8 0)
>      of {
>        False -> x_acFz;
>        True ->
>          rec_Bh
>            (- @Word8 $fNumWord8 x_acFz (fromInteger @Word8 $fNumWord8 1))
>      })
>   eta_B0

we then get the following panic:

> ghc: panic! (the 'impossible' happened)
>   (GHC version 9.2.1:
> 	refineFromInScope
>   InScope {wild_00}
>   eta_B0
>   Call stack:
>       CallStack (from HasCallStack):
>         callStackDoc, called at compiler/GHC/Utils/Panic.hs:181:37 in ghc:GHC.Utils.Panic
>         pprPanic, called at compiler/GHC/Core/Opt/Simplify/Env.hs:706:30 in ghc:GHC.Core.Opt.Simplify.Env

but the variable it's complaining about doesn't seem to be referenced ...